refactor: 계산, 매핑 클래스 분리
